Crud,Sql,Jpql

+++ Linealll/src/main/java/cundi/edu/co/service/impl/AutorServiceImpl.java
@@ -1,20 +1,18 @@
 package cundi.edu.co.service.impl;
 
 import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
 import cundi.edu.co.entity.Autor;
 import cundi.edu.co.exception.ArgumentRequiredException;
 import cundi.edu.co.exception.ConflicException;
 import cundi.edu.co.exception.ModelNotFoundException;
 import cundi.edu.co.repository.IAutorRepository;
 import cundi.edu.co.service.IAutorService;
+import javassist.NotFoundException;
 
 @Service
 public class AutorServiceImpl implements IAutorService {
@@ -44,12 +42,12 @@ public class AutorServiceImpl implements IAutorService {
 
 	@Override
 	public void eliminar(int idEntity) throws ModelNotFoundException {
-	//	if(validarExistenciaPorId(idEntity))
+		if(validarExistenciaPorId(idEntity)) {
 			this.repo.deleteById(idEntity);
-	//	else
-	//		throw new ModelNotFoundException("Estudiante no encontrado");
+		}else {
+			throw new ModelNotFoundException("Autor no encontrado");
+		}		
 	}
-	
 
 	@Override
 	public void crear(Autor entity) throws ConflicException {
@@ -65,5 +63,11 @@ public class AutorServiceImpl implements IAutorService {
 	private Boolean validarExistenciaPorId(int idAutor) {
 		return repo.existsById(idAutor);
 	}
+
+	@Override
+	public List<Autor> buscarCedula(String filtro) {
+		return repo.buscarCedula(filtro);
+	}
+	
 
 }
